requirements.txt
click==8.0.1
colorama==0.4.4
mypy-extensions==0.4.3
numpy==1.21.2
pandas==1.3.2
pathspec==0.9.0
platformdirs==2.2.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
six==1.16.0
tomli==1.2.1
typing-extensions==3.10.0.1
